AS-21: Rename Residence model file to ResidenceDetails
