+++ src/portal/residences/manager/providers/IStorageProvider.ts
@@ -2,7 +2,7 @@
  * Copyright 2024 Lou Amaya - All Rights Reserved
  */
 
-import { ResidenceDetails, ResidenceInfo } from "../../models/Residence"
+import { ResidenceDetails, ResidenceInfo } from "../../models/ResidenceDetails"
 
 export interface IStorageProvider {
     getResidenceList(): Promise<ResidenceInfo[]>
